# Optimize ordering of children in Flex layout