--- a/src/Controls/src/Core/Platform/Android/MultiPageFragmentStateAdapter.cs
+++ b/src/Controls/src/Core/Platform/Android/MultiPageFragmentStateAdapter.cs
@@ -7,11 +7,7 @@ using AndroidX.ViewPager2.Adapter;
 
 namespace Microsoft.Maui.Controls.Platform
 {
-	internal class MultiPageFragmentStateAdapter<[DynamicallyAccessedMembers(BindableProperty.DeclaringTypeMembers
-#if NET8_0 // IL2091
-	| BindableProperty.ReturnTypeMembers
-#endif
-	)] T> : FragmentStateAdapter where T : Page
+	internal class MultiPageFragmentStateAdapter<[DynamicallyAccessedMembers(BindableProperty.DeclaringTypeMembers | BindableProperty.ReturnTypeMembers)] T> : FragmentStateAdapter where T : Page
 	{
 		MultiPage<T> _page;
 		readonly IMauiContext _context;