--- a/src/Core/src/Layouts/Flex.cs
+++ b/src/Core/src/Layouts/Flex.cs
@@ -7,8 +7,8 @@
 //  - Stephane Delcroix (.NET port)
 //
 using System;
-using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Microsoft.Maui.Layouts.Flex
 {
@@ -990,27 +990,11 @@ namespace Microsoft.Maui.Layouts.Flex
 				ordered_indices = null;
 				if (item.ShouldOrderChildren && item.Count > 0)
 				{
-					var indices = new int[item.Count];
-					// Creating a list of item indices sorted using the children's `order'
-					// attribute values. We are using a simple insertion sort as we need
-					// stability (insertion order must be preserved) and cross-platform
-					// support. We should eventually switch to merge sort (or something
-					// else) if the number of items becomes significant enough.
-					for (int i = 0; i < item.Count; i++)
-					{
-						indices[i] = i;
-						for (int j = i; j > 0; j--)
-						{
-							int prev = indices[j - 1];
-							int curr = indices[j];
-							if (item[prev].Order <= item[curr].Order)
-							{
-								break;
-							}
-							indices[j - 1] = curr;
-							indices[j] = prev;
-						}
-					}
+					// The children need to be ordered using a *stable* sort by Order
+					var indices = item.Select((value, index) => new {value, index})
+									.OrderBy(x => x.value.Order)
+									.Select(x => x.index)
+									.ToArray();
 					ordered_indices = indices;
 				}