--- a/src/Core/src/Platform/iOS/MauiScrollView.cs
+++ b/src/Core/src/Platform/iOS/MauiScrollView.cs
@@ -68,9 +68,13 @@ namespace Microsoft.Maui.Platform
 		/// </summary>
 		bool _safeAreaInvalidated = true;
 
+		// Cached result of whether a parent MauiView is already applying safe area adjustments.
+		// Null means not yet determined. Invalidated when view hierarchy changes.
+		bool? _parentHandlesSafeArea;
+
 		/// <summary>
 		/// Flag indicating whether this scroll view should apply safe area adjustments to its content.
-		/// Only true when not nested in another scroll view and safe area is not empty.
+		/// Only true when not nested in another scroll view, no parent MauiView handles it, and safe area is not empty.
 		/// </summary>
 		bool _appliesSafeAreaAdjustments;
 
@@ -123,6 +127,20 @@ namespace Microsoft.Maui.Platform
 		}
 
 		/// <summary>
+		/// Checks if any ancestor MauiView is already applying safe area adjustments.
+		/// When a parent already handles safe area, this scroll view should not double-apply insets,
+		/// which would otherwise cause infinite layout cycles (#33595).
+		/// </summary>
+		bool IsParentHandlingSafeArea()
+		{
+			if (_parentHandlesSafeArea.HasValue)
+				return _parentHandlesSafeArea.Value;
+
+			_parentHandlesSafeArea = this.FindParent(x => x is MauiView mv && mv.AppliesSafeAreaAdjustments) is not null;
+			return _parentHandlesSafeArea.Value;
+		}
+
+		/// <summary>
 		/// Called by iOS when the adjusted content inset changes (e.g., when safe area changes).
 		/// This method invalidates the safe area and triggers a layout update if needed.
 		/// </summary>
@@ -148,8 +166,18 @@ namespace Microsoft.Maui.Platform
 		{
 			// Note: UIKit invokes LayoutSubviews right after this method
 			base.SafeAreaInsetsDidChange();
-
+			_parentHandlesSafeArea = null;
 			_safeAreaInvalidated = true;
+		}
+
+		/// <summary>
+		/// Directly invalidates this view's safe area, forcing re-evaluation on next layout pass.
+		/// </summary>
+		internal void InvalidateSafeArea()
+		{
+			_parentHandlesSafeArea = null;
+			_safeAreaInvalidated = true;
+			SetNeedsLayout();
 		}
 
 		/// <summary>
@@ -169,7 +197,7 @@ namespace Microsoft.Maui.Platform
 			{
 				return safeAreaPage.GetSafeAreaRegionsForEdge(edge);
 			}
-			
+
 			return SafeAreaRegions.None; // Default: edge-to-edge content
 		}
 
@@ -215,7 +243,7 @@ namespace Microsoft.Maui.Platform
 				// All edges have the same value, use built-in iOS behavior
 				// Cache the region value to avoid redundant comparisons
 				var region = leftRegion;
-				
+
 				ContentInsetAdjustmentBehavior = region switch
 				{
 					SafeAreaRegions.Default => UIScrollViewContentInsetAdjustmentBehavior.Automatic, // Default behavior
@@ -328,8 +356,9 @@ namespace Microsoft.Maui.Platform
 			//UpdateKeyboardSubscription();
 			// If nothing changed, we don't need to do anything
 
-			if (!UpdateContentInsetAdjustmentBehavior())
+			if (UpdateContentInsetAdjustmentBehavior())
 			{
+				// Edges changed - invalidate and force re-evaluation
 				InvalidateConstraintsCache();
 				_safeAreaInvalidated = true;
 			}
@@ -340,7 +369,7 @@ namespace Microsoft.Maui.Platform
 			}
 
 			// Mark the safe area as validated given that we're about to check it
-			_safeAreaInvalidated = true;
+			_safeAreaInvalidated = false;
 
 			var oldSafeArea = _safeArea;
 
@@ -356,7 +385,7 @@ namespace Microsoft.Maui.Platform
 				_safeArea = SystemAdjustedContentInset.ToSafeAreaInsets();
 
 			var oldApplyingSafeAreaAdjustments = _appliesSafeAreaAdjustments;
-			_appliesSafeAreaAdjustments = RespondsToSafeArea() && !_safeArea.IsEmpty;
+			_appliesSafeAreaAdjustments = !IsParentHandlingSafeArea() && RespondsToSafeArea() && !_safeArea.IsEmpty;
 
 			if (_systemAdjustedContentInset != SystemAdjustedContentInset)
 			{
@@ -370,9 +399,11 @@ namespace Microsoft.Maui.Platform
 				InvalidateConstraintsCache();
 			}
 
-			// Return whether the way safe area interacts with our view has changed
+			// Return whether the way safe area interacts with our view has changed.
+			// Compare at device-pixel resolution to filter sub-pixel noise from animations
+			// that would otherwise trigger infinite layout invalidation cycles (#32586, #33934).
 			return oldApplyingSafeAreaAdjustments == _appliesSafeAreaAdjustments &&
-				   (oldSafeArea == _safeArea || !_appliesSafeAreaAdjustments);
+				   (oldSafeArea.EqualsAtPixelLevel(_safeArea) || !_appliesSafeAreaAdjustments);
 		}
 
 		UIEdgeInsets SystemAdjustedContentInset
@@ -466,49 +497,40 @@ namespace Microsoft.Maui.Platform
 
 			contentSize = new Size(width, height);
 
-			// For Right-To-Left (RTL) layouts, we need to adjust the content arrangement and offset
-			// to ensure the content is correctly aligned and scrolled. This involves a second layout
-			// arrangement with an adjusted starting point and recalculating the content offset.
-			if (_previousEffectiveUserInterfaceLayoutDirection != EffectiveUserInterfaceLayoutDirection)
+			bool isDirectionChange = _previousEffectiveUserInterfaceLayoutDirection != EffectiveUserInterfaceLayoutDirection;
+
+			// For Right-To-Left (RTL) layouts, iOS natively handles visual mirroring via
+			// SemanticContentAttribute.ForceRightToLeft. Content should remain at normal (0,0) coordinates.
+			// We only set ContentOffset to position the scroll at the RTL "start" (maximum horizontal offset).
+			// Content at negative X coordinates would be outside the scrollable range and unreachable.
+			if (isDirectionChange)
 			{
-				// In mac platform, Scrollbar is not updated based on FlowDirection, so resetting the scroll indicators
-				// It's a native limitation; to maintain platform consistency, a hack fix is applied to show the Scrollbar based on the FlowDirection.
-				if (OperatingSystem.IsMacCatalyst() && _previousEffectiveUserInterfaceLayoutDirection is not null)
-				{
-					bool showsVertical = ShowsVerticalScrollIndicator;
-					bool showsHorizontal = ShowsHorizontalScrollIndicator;
-
-					ShowsVerticalScrollIndicator = false;
-					ShowsHorizontalScrollIndicator = false;
-
-					ShowsVerticalScrollIndicator = showsVertical;
-					ShowsHorizontalScrollIndicator = showsHorizontal;
-				}
-
 				if (EffectiveUserInterfaceLayoutDirection == UIUserInterfaceLayoutDirection.RightToLeft)
 				{
-					var horizontalOffset = contentSize.Width - bounds.Width;
-					
-					if (SystemAdjustedContentInset == UIEdgeInsets.Zero || ContentInsetAdjustmentBehavior == UIScrollViewContentInsetAdjustmentBehavior.Never)
+					// In mac platform, Scrollbar is not updated based on FlowDirection, so resetting the scroll indicators.
+					// It's a native limitation; to maintain platform consistency, a hack fix is applied to show the Scrollbar based on the FlowDirection.
+					if (OperatingSystem.IsMacCatalyst() && _previousEffectiveUserInterfaceLayoutDirection is not null)
 					{
-						CrossPlatformLayout?.CrossPlatformArrange(new Rect(new Point(-horizontalOffset, 0), bounds.Size.ToSize()));
-					}
-					else
-					{
-						CrossPlatformLayout?.CrossPlatformArrange(new Rect(new Point(-horizontalOffset, 0), bounds.Size.ToSize()));
-					}
-					
-					ContentOffset = new CGPoint(horizontalOffset, 0);
+						bool showsVertical = ShowsVerticalScrollIndicator;
+						bool showsHorizontal = ShowsHorizontalScrollIndicator;
 
+						ShowsVerticalScrollIndicator = false;
+						ShowsHorizontalScrollIndicator = false;
+
+						ShowsVerticalScrollIndicator = showsVertical;
+						ShowsHorizontalScrollIndicator = showsHorizontal;
+					}
+
+					var horizontalOffset = contentSize.Width - bounds.Width;
+					ContentOffset = new CGPoint(horizontalOffset, 0);
 				}
-				else if(_previousEffectiveUserInterfaceLayoutDirection is not null)
+				else if (_previousEffectiveUserInterfaceLayoutDirection is not null)
 				{
 					ContentOffset = new CGPoint(0, ContentOffset.Y);
 				}
 			}
 
-			// When switching between LTR and RTL, we need to re-arrange and offset content exactly once
-			// to avoid cumulative shifts or incorrect offsets on subsequent layouts.
+			// Track the current direction so we can detect future changes.
 			_previousEffectiveUserInterfaceLayoutDirection = EffectiveUserInterfaceLayoutDirection;
 
 			return contentSize;
@@ -548,7 +570,7 @@ namespace Microsoft.Maui.Platform
 		/// </summary>
 		/// <param name="size">The available size constraints.</param>
 		/// <returns>The size that fits within the constraints.</returns>
-		
+
 		public override CGSize SizeThatFits(CGSize size)
 		{
 			if (CrossPlatformLayout is null)
@@ -671,6 +693,7 @@ namespace Microsoft.Maui.Platform
 
 			// Clear cached scroll view descendant status since the view hierarchy may have changed
 			_scrollViewDescendant = null;
+			_parentHandlesSafeArea = null;
 
 			// Mark safe area as invalidated since moving to a new window may change safe area
 			_safeAreaInvalidated = true;