--- a/src/Core/src/Platform/iOS/MauiView.cs
+++ b/src/Core/src/Platform/iOS/MauiView.cs
@@ -68,6 +68,10 @@ namespace Microsoft.Maui.Platform
 		// otherwise, false. Null means not yet determined.
 		bool? _scrollViewDescendant;
 
+		// Cached result of whether a parent MauiView is already handling safe area.
+		// Null means not yet determined. Invalidated when view hierarchy changes.
+		bool? _parentHandlesSafeArea;
+
 		// Keyboard tracking
 		CGRect _keyboardFrame = CGRect.Empty;
 		bool _isKeyboardShowing;
@@ -383,6 +387,42 @@ namespace Microsoft.Maui.Platform
 
 
 		/// <summary>
+		/// Returns whether this view is currently applying safe area adjustments to its layout.
+		/// Used by descendant views to avoid double-applying safe area when a parent already handles it.
+		/// </summary>
+		internal bool AppliesSafeAreaAdjustments => _appliesSafeAreaAdjustments;
+
+		/// <summary>
+		/// Checks if any ancestor MauiView is already applying safe area adjustments for the same edges
+		/// that this view handles. When a parent already handles a specific safe area edge, this view
+		/// should not double-apply insets for that edge — but it may still handle OTHER edges independently.
+		/// This prevents double-padding when parent and child handle the same edges (#33595, #32586),
+		/// while allowing parent and child to handle DIFFERENT edges without conflict (#28986).
+		/// </summary>
+		bool IsParentHandlingSafeArea()
+		{
+			if (_parentHandlesSafeArea.HasValue)
+				return _parentHandlesSafeArea.Value;
+
+			// Check if any ancestor MauiView handles any of the SAME edges we handle.
+			// Edge-aware check: parent handling only TOP doesn't block child handling BOTTOM.
+			_parentHandlesSafeArea = this.FindParent(x =>
+			{
+				if (x is not MauiView mv || !mv._appliesSafeAreaAdjustments)
+					return false;
+				// Return true only if parent handles any edge that this view also handles
+				for (int edge = 0; edge < 4; edge++)
+				{
+					if (GetSafeAreaRegionForEdge(edge) != SafeAreaRegions.None &&
+						mv.GetSafeAreaRegionForEdge(edge) != SafeAreaRegions.None)
+						return true;
+				}
+				return false;
+			}) is not null;
+			return _parentHandlesSafeArea.Value;
+		}
+
+		/// <summary>
 		/// Checks if the current measure information is still valid for the given constraints.
 		/// This optimization avoids redundant measure operations when constraints haven't changed.
 		/// </summary>
@@ -610,11 +650,13 @@ namespace Microsoft.Maui.Platform
 			_safeArea = GetAdjustedSafeAreaInsets();
 
 			var oldApplyingSafeAreaAdjustments = _appliesSafeAreaAdjustments;
-			_appliesSafeAreaAdjustments = RespondsToSafeArea() && !_safeArea.IsEmpty;
+			_appliesSafeAreaAdjustments = !IsParentHandlingSafeArea() && RespondsToSafeArea() && !_safeArea.IsEmpty;
 
-			// Return whether the way safe area interacts with our view has changed
+			// Return whether the way safe area interacts with our view has changed.
+			// Compare at device-pixel resolution to filter sub-pixel noise from animations
+			// that would otherwise trigger infinite layout invalidation cycles (#32586, #33934).
 			return oldApplyingSafeAreaAdjustments == _appliesSafeAreaAdjustments &&
-				   (oldSafeArea == _safeArea || !_appliesSafeAreaAdjustments);
+				   (oldSafeArea.EqualsAtPixelLevel(_safeArea) || !_appliesSafeAreaAdjustments);
 		}
 
 		/// <summary>
@@ -702,7 +744,18 @@ namespace Microsoft.Maui.Platform
 		public override void SafeAreaInsetsDidChange()
 		{
 			_safeAreaInvalidated = true;
+			_parentHandlesSafeArea = null;
 			base.SafeAreaInsetsDidChange();
+		}
+
+		/// <summary>
+		/// Directly invalidates this view's safe area, forcing re-evaluation on next layout pass.
+		/// </summary>
+		internal void InvalidateSafeArea()
+		{
+			_safeAreaInvalidated = true;
+			_parentHandlesSafeArea = null;
+			SetNeedsLayout();
 		}
 
 		/// <summary>
@@ -714,6 +767,7 @@ namespace Microsoft.Maui.Platform
 			base.MovedToWindow();
 
 			_scrollViewDescendant = null;
+			_parentHandlesSafeArea = null;
 
 			// Notify any subscribers that this view has been moved to a window
 			_movedToWindow?.Invoke(this, EventArgs.Empty);