--- a/src/Core/src/Platform/iOS/SafeAreaPadding.cs
+++ b/src/Core/src/Platform/iOS/SafeAreaPadding.cs
@@ -28,6 +28,23 @@ internal readonly record struct SafeAreaPadding(double Left, double Right, doubl
 
 	public CGRect ToCGRect() =>
 		new((nfloat)Top, (nfloat)Left, (nfloat)Bottom, (nfloat)Right);
+
+	/// <summary>
+	/// Compares two SafeAreaPadding values at device-pixel resolution.
+	/// Sub-pixel differences (e.g., 0.001pt from animation noise) that map to the same
+	/// physical pixel are treated as equal, preventing unnecessary layout invalidation cycles.
+	/// </summary>
+	public bool EqualsAtPixelLevel(SafeAreaPadding other)
+	{
+		var scale = (double)UIScreen.MainScreen.Scale;
+		return RoundToPixel(Left, scale) == RoundToPixel(other.Left, scale)
+			&& RoundToPixel(Right, scale) == RoundToPixel(other.Right, scale)
+			&& RoundToPixel(Top, scale) == RoundToPixel(other.Top, scale)
+			&& RoundToPixel(Bottom, scale) == RoundToPixel(other.Bottom, scale);
+	}
+
+	static double RoundToPixel(double value, double scale)
+		=> Math.Round(value * scale, MidpointRounding.AwayFromZero);
 }
 
 internal static class SafeAreaInsetsExtensions