… .github/copilot-instructions.md
… .github/instructions/safe-area-ios.instructions.md
… .github/instructions/uitests.instructions.md
… eng/pipelines/ci-uitests.yml
… eng/pipelines/ci.yml
… src/Controls/tests/TestCases.Android.Tests/snapshots/android/LabelNotTruncatedWithMaxLines.png
… src/Controls/tests/TestCases.HostApp/Issues/Issue28986_ParentChildTest.xaml
… src/Controls/tests/TestCases.HostApp/Issues/Issue28986_ParentChildTest.xaml.cs
… src/Controls/tests/TestCases.HostApp/Issues/Issue32586.cs
… src/Controls/tests/TestCases.HostApp/Issues/Issue33595.cs
… src/Controls/tests/TestCases.HostApp/Issues/Issue33934.xaml
… src/Controls/tests/TestCases.HostApp/Issues/Issue33934.xaml.cs
… src/Controls/tests/TestCases.HostApp/Issues/Issue33934DialogBase.cs
… src/Controls/tests/TestCases.HostApp/Issues/Issue33934DialogPage.xaml
… src/Controls/tests/TestCases.HostApp/Issues/Issue33934DialogPage.xaml.cs
… src/Controls/tests/TestCases.HostApp/Issues/Issue33934ViewPage.cs
… src/Controls/tests/TestCases.HostApp/Issues/Issue34120.cs
… src/Controls/tests/TestCases.HostApp/Issues/ViewModelBase.cs
… src/Controls/tests/TestCases.Mac.Tests/snapshots/mac/LabelNotTruncatedWithMaxLines.png
… src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue28986_ParentChildTest.cs
… src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue32586.cs
… src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue33595.cs
… src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue33934.cs
… src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue34120.cs
… src/Controls/tests/TestCases.WinUI.Tests/snapshots/windows/LabelNotTruncatedWithMaxLines.png
… src/Controls/tests/TestCases.iOS.Tests/snapshots/ios-26/LabelNotTruncatedWithMaxLines.png
… src/Controls/tests/TestCases.iOS.Tests/snapshots/ios/LabelNotTruncatedWithMaxLines.png
… src/TestUtils/src/Microsoft.Maui.IntegrationTests/AOTTemplateTest.cs