This had a TODO comment to use mergesort, but LINQ offers a highly optimized quicksort that would be just as good.
 
Before:

| Method                    | Mean     | Error     | StdDev    | Median   | Gen0     | Completed Work Items | Lock Contentions | Gen1     | Gen2    | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|---------------------:|-----------------:|---------:|--------:|----------:|
| LayoutLotsOfItemsWithWrap | 6.331 ms | 0.7514 ms | 2.2156 ms | 4.612 ms | 171.8750 |                    - |                - | 140.6250 | 31.2500 |   1.57 MB |
| LayoutLotsOfItemsNoWrap   | 4.432 ms | 0.0798 ms | 0.0747 ms | 4.402 ms | 171.8750 |                    - |                - | 140.6250 | 31.2500 |   1.57 MB |

After:

| Method                    | Mean     | Error     | StdDev    | Median   | Gen0     | Completed Work Items | Lock Contentions | Gen1     | Gen2    | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|---------------------:|-----------------:|---------:|--------:|----------:|
| LayoutLotsOfItemsWithWrap | 4.513 ms | 0.2016 ms | 0.5312 ms | 4.336 ms | 171.8750 |                    - |                - | 132.8125 | 39.0625 |   1.56 MB |
| LayoutLotsOfItemsNoWrap   | 4.479 ms | 0.0850 ms | 0.0754 ms | 4.470 ms | 171.8750 |                    - |                - | 132.8125 | 39.0625 |   1.56 MB |