@mattleibow 2024-04-23
> Another idea: Having a struct for value+index might lead to better perf. I'm not too sure what { value, index } will create, if a struct or a class behind the scenes.

@MartyIX can you use a value tuple? `(value, index)` Does that make a difference with allocations or something?