@jonathanpeppers 2025-06-09
@mattleibow I don't know if I trust the AI comments, because it hallucinated invalid code:

```csharp
var indices = item.Select((x, i) => new { x, i }) // Why using new?
                  .OrderBy(x => x.x.Order) // Ok, maybe .x.x is Psuedocode?
                  .Select(x => x.i) // But then if this is an int, what is .i?
                  .ToArray();
```

Should we just prove out the new code with a unit test or two?

We could check `Children` with the same `Order` are stable sorted?